requirements.txt
streamlit==1.31.0
groq==0.4.1
langchain==0.1.10
langchain-community==0.0.25
langchain-groq==0.0.1
faiss-cpu==1.9.0.post1
pypdf==4.0.1
python-dotenv==1.0.1
supabase==2.3.4
sentence-transformers==2.3.1
validators==0.22.0
python-dateutil==2.8.2